requirements.txt
streamlit
web3
requests
python-dotenv
matplotlib
pandas
